+++ ethereum/interfaces/IWormholeAdapterPool.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.0;
 interface IWormholeAdapterPool {
     function dolaChainId() external view returns (uint16);
 
-    function getNonce() external view returns (uint32);
+    function getNonce() external view returns (uint64);
 
     function sendDeposit(
         address pool,
